Renamed test modules to tests
