+++ src/encoder/tests.rs
